Process access_token from successful OAuth login
The response comes back in bytes, which needs to
be converted into a string. But that also comes
in a hodge-podge of JSON values, of which only a
few are really needed. Not wanting to create a
struct for the whole thing, using a map should
be good enough for now.