Use more log.Fatal to reduce all the indents
Instead of all the if/else controls, the use of just
log.Fatal to spit out the error and moving on cuts down
on how far more indentation needs to happen.

Adds use of defer to close out the HTTP request from
the fetch. Not sure it needs to happen, but reading
so far suggests it is a clean thing to do.