Leave created records sitting around as default
For the sake of quickly populating a bunch of records, the hohum
would be to have these newly created records sit around for a
little bit.

The need to purge them right after is probably the second-tier
use case of just seeing if the delete still works in a test run.